requirements.txt
flask==3.1.2
pandas==3.0.0
python-dotenv==1.2.1
apscheduler==3.10.4
twilio==9.0.4
flask-login==0.6.3
werkzeug==3.1.3
gunicorn==21.2.0
